requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Google Workspace (Sheets, Docs, Drive, Gmail)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# File attachments
pypdf>=4.0.0
python-docx>=1.0.0

# Voice (speech-to-text + audio recording)
openai-whisper>=20230918
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
